--- Assets/Prefabs/Terrain Generation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5ad09d045b1b5d9439da6a646a8b4781
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
