--- Assets/Prefabs/Terrain Generation/Level Tile.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7d75268bdd1c64339ace1846f46d13be
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
