+++ Assets/Scripts/WorldGeneration/RiverGeneration.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: c5a6a551e261a4491975d98c53c3455c
+guid: c6ea9aef9f6d1dd4586c8af63de8dcb3
+timeCreated: 1507637110
+licenseType: Free
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
